+++ database/migrations/2024_05_25_172019_create_comandas_table.php
@@ -14,12 +14,12 @@ return new class extends Migration
         Schema::create('comandas', function (Blueprint $table) {
             $table->id();
             $table->string('nome', 25)->nullable();
-            $table->decimal('total', 8, 2)->nullable();
-            $table->decimal('lucro', 8, 2)->nullable();
+            $table->float('total', 8, 2)->nullable();
+            $table->float('lucro', 8, 2)->nullable();
             $table->unsignedBigInteger('forma_pagamento_id')->nullable();
-            $table->decimal('taxa', 4, 2)->nullable();
-            $table->decimal('dinheiro', 8, 2)->nullable();
-            $table->decimal('troco', 8, 2)->nullable();
+            $table->float('taxa', 4, 2)->nullable();
+            $table->float('dinheiro', 8, 2)->nullable();
+            $table->float('troco', 8, 2)->nullable();
             $table->enum('status', ['a','f'])->default('a');
             $table->datetime('data_abertura')->nullable();
             $table->datetime('data_fechamento')->nullable();
