+++ database/migrations/2024_05_27_185712_create_pedidos_table.php
@@ -14,14 +14,14 @@ return new class extends Migration
         Schema::create('pedidos', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id')->nullable();
-            $table->decimal('total', 10, 2)->nullable();
-            $table->decimal('lucro', 10, 2)->nullable();
+            $table->float('total', 10, 2)->nullable();
+            $table->float('lucro', 10, 2)->nullable();
             $table->unsignedBigInteger('forma_pagamento_id')->nullable();
             $table->integer('pontos')->nullable();
             $table->integer('pontos_troca')->nullable();
-            $table->decimal('dinheiro', 10, 2)->nullable();
-            $table->decimal('troco', 10, 2)->nullable();
-            $table->decimal('frete', 10, 2)->nullable();
+            $table->float('dinheiro', 10, 2)->nullable();
+            $table->float('troco', 10, 2)->nullable();
+            $table->float('frete', 10, 2)->nullable();
             $table->string('novo_endereco', 100)->nullable();
             $table->string('codigo', 4)->nullable();
             $table->enum('status', ['n','s', 'ac','e'])->nullable()->default('n');
