+++ database/migrations/2024_06_10_180144_create_fluxo_caixas_table.php
@@ -14,9 +14,9 @@ return new class extends Migration
         Schema::create('fluxo_caixas', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('caixa_id')->nullable(false);
-            $table->decimal('venda',10,2)->nullable();
-            $table->decimal('dinheiro',10,2)->nullable();
-            $table->decimal('troco',10,2)->nullable();
+            $table->float('venda',10,2)->nullable();
+            $table->float('dinheiro',10,2)->nullable();
+            $table->float('troco',10,2)->nullable();
             $table->datetime('data')->nullable();
 
             $table->foreign('caixa_id')->references('id')->on('caixas')->onDelete('cascade');
